noop split between ray and python launchers
