+++ data-processing-lib/doc/testing-transforms.md
@@ -34,7 +34,7 @@ from typing import Tuple
 
 import pyarrow as pa
 from data_processing.test_support import AbstractTransformTest
-from noop_transform import NOOPTransform
+from noop_transform_ray import NOOPTransform
 
 # Define the test input and expected outputs
 table = pa.Table.from_pydict({"name": pa.array(["Tom"]), "age": pa.array([23])})
@@ -44,13 +44,13 @@ expected_metadata_list = [{"nfiles": 1, "nrows": 1}, {}]  # transform() result  
 
 class TestNOOPTransform(AbstractTransformTest):
 
-  # Define the method that provides the test fixtures to the test from the super class.
-  def get_test_transform_fixtures(self) -> list[Tuple]:
-    fixtures = [
-      (NOOPTransform({"sleep": 0}), [table], [expected_table], expected_metadata_list),
-      (NOOPTransform({"sleep": 1}), [table], [expected_table], expected_metadata_list),
-    ]
-    return fixtures
+    # Define the method that provides the test fixtures to the test from the super class.
+    def get_test_transform_fixtures(self) -> list[Tuple]:
+        fixtures = [
+            (NOOPTransform({"sleep": 0}), [table], [expected_table], expected_metadata_list),
+            (NOOPTransform({"sleep": 1}), [table], [expected_table], expected_metadata_list),
+        ]
+        return fixtures
 ```
 In the above we use the `NOOPTransform` to process the single input `table`, to produce
 the expected table `expected_table` and list of metadata in `expected_metadata_list`,
