+++ transforms/.make.transforms
@@ -215,8 +215,13 @@ conventions:
 
 .PHONY: .transforms.run-cli-ray-sample 
 .transforms.run-cli-ray-sample:
-	$(MAKE) RUN_FILE=$(TRANSFORM_NAME)_transform.py \
+	$(MAKE) RUN_FILE=$(TRANSFORM_NAME)_transform_ray.py \
 		RUN_ARGS="--run_locally True --data_local_config \"{ 'input_folder' : '../test-data/input', 'output_folder' : '../output'}\" " .transforms.run-src-file
+
+.PHONY: .transforms.run-cli-python-sample 
+.transforms.run-cli-python-sample:
+	$(MAKE) RUN_FILE=$(TRANSFORM_NAME)_transform.py \
+		RUN_ARGS=" --data_local_config \"{ 'input_folder' : '../test-data/input', 'output_folder' : '../output'}\" " .transforms.run-src-file
 
 # Expects RUN_FILE
 .PHONY: .transforms.run-src-file
