+++ transforms/universal/noop/src/noop_transform.py
@@ -15,7 +15,7 @@ from argparse import ArgumentParser, Namespace
 from typing import Any
 
 import pyarrow as pa
-from data_processing.ray import TableTransformConfigurationRay, TransformLauncherRay
+from data_processing.pure_python import TransformLauncher
 from data_processing.transform import AbstractTableTransform, TransformConfiguration
 from data_processing.utils import CLIArgumentProvider, get_logger
 
@@ -115,21 +115,6 @@ class NOOPTransformConfigurationBase:
         return True
 
 
-class NOOPTransformConfigurationRay(TableTransformConfigurationRay):
-    def __init__(self):
-        super().__init__(name=short_name, transform_class=NOOPTransform, remove_from_metadata=[pwd_key])
-        self.base = NOOPTransformConfigurationBase()
-
-    def add_input_params(self, parser: ArgumentParser) -> None:
-        return self.base.add_input_params(parser=parser)
-
-    def apply_input_params(self, args: Namespace) -> bool:
-        is_valid = self.base.apply_input_params(args=args)
-        if is_valid:
-            self.params = self.base.params
-        return is_valid
-
-
 class NOOPTransformConfigurationPython(TransformConfiguration):
     def __init__(self):
         super().__init__(name=short_name, transform_class=NOOPTransform, remove_from_metadata=[pwd_key])
@@ -146,6 +131,6 @@ class NOOPTransformConfigurationPython(TransformConfiguration):
 
 
 if __name__ == "__main__":
-    launcher = TransformLauncherRay(transform_runtime_config=NOOPTransformConfigurationRay())
+    launcher = TransformLauncher(transform_runtime_config=NOOPTransformConfigurationPython())
     logger.info("Launching noop transform")
     launcher.launch()
